backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
tensorflow==2.15.0
numpy==1.26.4
opencv-python==4.9.0.80
reportlab==4.0.9
pillow==10.2.0
pillow-heif==0.15.0
groq==0.11.0
gunicorn==21.2.0
